+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   * tsm_storage_pool_utilized_percent
   * tsm_volume_utilized_percent
 * Remove reason and servername label from tsm_status
+* Make tsm_db_buffer_hit_ratio and tsm_db_pkg_hit_ratio a ratio between 0.0-1.0
+* Rename tsm_db_buffer_total_requests to tsm_db_buffer_requests_total
 
 ### Changes